+++ examples/cc_asg/outputs.tf
@@ -25,7 +25,7 @@ Forwarding/Service VPC Network:
 ${module.cc_vm.instance_template_forwarding_vpc}
 
 Internal Load Balancer IP:
-${module.ilb.next_hop_ilb_ip_address}
+${local.ilb_ip}
 
 Public Load Balancer Frontend IP:
 ${local.glb_ip}
@@ -65,7 +65,8 @@ TB
 }
 
 locals {
-  glb_ip = (one(module.glb[*].next_hop_glb_ip_address) == null) ? "" : one(module.glb[*].next_hop_glb_ip_address)
+  ilb_ip = module.ilb.next_hop_ilb_ip_address
+  glb_ip = (one(module.glb[*].glb_frontend_ip_address) == null) ? "" : one(module.glb[*].glb_frontend_ip_address)
 }
 
 output "testbedconfig" {