+++ examples/cc_ilb/main.tf
@@ -80,19 +80,19 @@ module "network" {
 ################################################################################
 # Create the user_data file with necessary bootstrap variables for Cloud Connector registration
 locals {
-  GLB_VIP = (var.glb_deploy == true) ? "\"glb_vip\": \"${module.glb[0].glb_ip_address}\"," : ""
-  ILB_VIP = (var.ilb_enabled == true) ? "\"lb_vip\": \"${module.ilb[0].ilb_ip_address}\"," : ""
+  glb_vip = var.glb_deploy ? "\"glb_vip\": \"${module.glb[0].glb_ip_address}\"," : ""
+  ilb_vip = var.ilb_enabled ? "\"lb_vip\": \"${module.ilb[0].ilb_ip_address}\"," : ""
   # Populate potential locals map with HCP Vault variables
   hcpuserdata = <<USERDATA
 {
-  ${local.GLB_VIP}
+  ${local.glb_vip}
   "cc_url": "${var.cc_vm_prov_url}",
   "http_probe_port": ${var.http_probe_port},
   "hcp_vault_addr": "${var.hcp_vault_address}",
   "hcp_vault_secret_path": "${var.hcp_vault_secret_path}",
   "hcp_vault_role_name": "${var.hcp_vault_role_name}",
   "hcp_gcp_auth_role_type": "${var.hcp_gcp_auth_role_type}",
-  ${local.ILB_VIP}
+  ${local.ilb_vip}
   "gcp_service_account": "${module.iam_service_account.service_account}"
 }
 USERDATA
@@ -100,11 +100,11 @@ USERDATA
   # Populate potential local map with default GCP Secret Manager
   userdata = <<USERDATA
 {
-  ${local.GLB_VIP}
+  ${local.glb_vip}
   "cc_url": "${var.cc_vm_prov_url}",
   "secret_name": "${var.secret_name}",
   "http_probe_port": ${var.http_probe_port},
-  ${local.ILB_VIP}
+  ${local.ilb_vip}
   "gcp_service_account": "${module.iam_service_account.service_account}"
 }
 USERDATA
@@ -205,7 +205,7 @@ module "ilb" {
   source                      = "../../modules/terraform-zscc-ilb-gcp"
   vpc_network                 = module.network.service_vpc_network
   project                     = var.project
-  project_host                = var.project_host #optional
+  project_host                = var.project_host
   region                      = var.region
   instance_groups             = local.instance_groups_list
   vpc_subnetwork_ccvm_service = module.network.service_subnet
@@ -225,10 +225,10 @@ module "ilb" {
 
 module "glb" {
   source                      = "../../modules/terraform-zscc-glb-gcp"
-  count                       = (var.glb_deploy == true) ? 1 : 0
+  count                       = var.glb_deploy ? 1 : 0
   vpc_network                 = module.network.service_vpc_network
   project                     = var.project
-  project_host                = var.project_host #optional
+  project_host                = var.project_host
   region                      = var.region
   instance_groups             = local.instance_groups_list
   vpc_subnetwork_ccvm_service = module.network.service_subnet
@@ -251,12 +251,12 @@ module "glb" {
 ################################################################################
 module "cloud_dns" {
   source         = "../../modules/terraform-zscc-cloud-dns-gcp"
-  count          = var.zpa_enabled == true ? 1 : 0
+  count          = var.zpa_enabled ? 1 : 0
   name_prefix    = var.name_prefix
   resource_tag   = random_string.suffix.result
   vpc_networks   = [module.network.service_vpc_network]
   domain_names   = var.domain_names
   target_address = [module.ilb[0].ilb_ip_address]
   project        = var.project
-  project_host   = var.project_host #optional
+  project_host   = var.project_host
 }