+++ modules/terraform-zscc-glb-gcp/outputs.tf
@@ -1,10 +1,10 @@
-output "next_hop_glb_ip_address" {
-  description = "ILB front end IP address"
+output "glb_frontend_ip_address" {
+  description = "GLB frontend IP address"
   value       = google_compute_forwarding_rule.glb_forwarding.ip_address
 }
 
 output "next_hop_glb" {
-  description = "ID for ILB IP"
+  description = "ID for GLB forwarding rule"
   value       = google_compute_forwarding_rule.glb_forwarding.self_link
 }
 
@@ -14,6 +14,6 @@ output "glb_ip_address" {
 }
 
 output "glb_ip_address_link" {
-  description = "ID for ILB designated compute address"
+  description = "Self-link for GLB reserved compute address"
   value       = google_compute_address.glb_ip_address.self_link
 }